requirements.txt
streamlit
transformers
langchain-huggingface
faiss-cpu
nltk
reportlab
sentence-transformers
torch
langchain
